# fl_backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v0/schiedsrichter": {
      "get": {
        "summary": "List Schiedsrichter",
        "description": "List referees; deactivated ones are soft-deleted rather than removed, so historical matches stay resolvable.",
        "operationId": "get_schiedsrichter_api_v0_schiedsrichter_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "default_payment",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Default Payment"
            }
          },
          {
            "name": "include_inactive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Inactive"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1024,
              "minimum": 1,
              "default": 1024,
              "title": "Limit"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "name",
                "default_payment"
              ],
              "type": "string",
              "default": "name",
              "title": "Sort By"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "asc",
              "title": "Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FLSchiedsrichterListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a Schiedsrichter",
        "description": "Create a referee. `inactive_since` is set to null here and is not part of the payload.",
        "operationId": "post_schiedsrichter_api_v0_schiedsrichter_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FLPostSchiedsrichterPayload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FLPostSchiedsrichterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v0/schiedsrichter/{schiedsrichter_id}": {
      "get": {
        "summary": "One Schiedsrichter",
        "description": "Return one referee by their id, deactivated ones included -- a historical match references them by id.",
        "operationId": "get_schiedsrichter_by_id_api_v0_schiedsrichter__schiedsrichter_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "schiedsrichter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Schiedsrichter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FLSchiedsrichterSingleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update a Schiedsrichter and fan the change out",
        "description": "Update a referee, then update the embedded name on every Spiel that uses them.\n\nOnly the name. `payment` is NOT propagated: the fee on a match is what was agreed for it.",
        "operationId": "patch_schiedsrichter_api_v0_schiedsrichter__schiedsrichter_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "schiedsrichter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Schiedsrichter Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FLPatchSchiedsrichterPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FLPatchSchiedsrichterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Deactivate a Schiedsrichter (soft delete)",
        "description": "Deactivate a referee. SOFT, for the same reason as venues: matches embed a copy.",
        "operationId": "delete_schiedsrichter_api_v0_schiedsrichter__schiedsrichter_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "schiedsrichter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Schiedsrichter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FLSchiedsrichterWriteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FLKontakt": {
        "properties": {
          "telefon": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^([+]?[ 0-9\\-().]{3,20})$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefon"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "telefon",
          "email"
        ],
        "title": "FLKontakt"
      },
...
      "FLPatchSchiedsrichterPayload": {
        "properties": {
          "kontakt": {
            "$ref": "#/components/schemas/FLKontakt"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "pattern": "^\\p{L}[\\p{L}\\-' ]*$",
            "title": "Name"
          },
          "schule": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schule"
          },
          "default_payment": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Default Payment"
          }
        },
        "type": "object",
        "required": [
          "kontakt",
          "name",
          "schule",
          "default_payment"
        ],
        "title": "FLPatchSchiedsrichterPayload"
      },
      "FLPatchSchiedsrichterResponse": {
        "properties": {
          "acknowledged": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "title": "Acknowledged",
            "default": 1
          },
          "updated_document": {
            "$ref": "#/components/schemas/FLSchiedsrichter"
          }
        },
        "type": "object",
        "required": [
          "updated_document"
        ],
        "title": "FLPatchSchiedsrichterResponse"
      },
...
      "FLPostSchiedsrichterPayload": {
        "properties": {
          "kontakt": {
            "$ref": "#/components/schemas/FLKontakt"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "pattern": "^\\p{L}[\\p{L}\\-' ]*$",
            "title": "Name"
          },
          "schule": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schule"
          },
          "default_payment": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Default Payment"
          }
        },
        "type": "object",
        "required": [
          "kontakt",
          "name",
          "schule",
          "default_payment"
        ],
        "title": "FLPostSchiedsrichterPayload"
      },
      "FLPostSchiedsrichterResponse": {
        "properties": {
          "acknowledged": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "title": "Acknowledged",
            "default": 1
          },
          "created_id": {
            "type": "string",
            "title": "Created Id"
          }
        },
        "type": "object",
        "required": [
          "created_id"
        ],
        "title": "FLPostSchiedsrichterResponse"
      },
...
      "FLSchiedsrichter": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "schule": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schule"
          },
          "default_payment": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Default Payment"
          },
          "kontakt": {
            "$ref": "#/components/schemas/FLKontakt"
          },
          "inactive_since": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9]{4}-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01])$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inactive Since"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "schule",
          "default_payment",
          "kontakt",
          "inactive_since"
        ],
        "title": "FLSchiedsrichter"
      },
      "FLSchiedsrichterListResponse": {
        "properties": {
          "acknowledged": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "title": "Acknowledged",
            "default": 1
          },
          "schiedsrichter": {
            "items": {
              "$ref": "#/components/schemas/FLSchiedsrichter"
            },
            "type": "array",
            "title": "Schiedsrichter"
          }
        },
        "type": "object",
        "required": [
          "schiedsrichter"
        ],
        "title": "FLSchiedsrichterListResponse"
      },
      "FLSchiedsrichterSingleResponse": {
        "properties": {
          "acknowledged": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "title": "Acknowledged",
            "default": 1
          },
          "schiedsrichter": {
            "$ref": "#/components/schemas/FLSchiedsrichter"
          }
        },
        "type": "object",
        "required": [
          "schiedsrichter"
        ],
        "title": "FLSchiedsrichterSingleResponse"
      },
      "FLSchiedsrichterWriteResponse": {
        "properties": {
          "acknowledged": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "title": "Acknowledged",
            "default": 1
          },
          "updated_document": {
            "$ref": "#/components/schemas/FLSchiedsrichter"
          }
        },
        "type": "object",
        "required": [
          "updated_document"
        ],
        "title": "FLSchiedsrichterWriteResponse",
        "description": "Shared by delete and reactivate — both answer with the referee as they now stand."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}